BonjourTables now sets the main IP when selecting
Selecting an item from the list in the BonjourTables fragment sets the
currently active IP to that selected one. The first item which is
instructions will set it to the currently active IP address as a
default.
MainActivity has a new preference change listener in case we go to
settings from BonjourTables and change the current IP then try to select
an item from the list.
Settings has had the Bonjour Table selection removed because the
fragment is fully featured now.